# api/docs/swagger.yaml
definitions:
...
  models.CategoryPrduct:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
      type_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateProduct:
    properties:
      branch_id:
        type: string
      category_id:
        type: string
      description:
        type: string
      name:
        type: string
      photo:
        type: string
      price:
        type: number
    type: object
  models.CreateUser:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  models.Empty:
    type: object
...
  models.GetListProductResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/models.Productlist'
        type: array
      count:
        type: integer
    type: object
  models.GetListUserResponse:
    properties:
      count:
        type: integer
      user:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.Product:
    properties:
      branch_id:
        type: string
      category_id:
        $ref: '#/definitions/models.CategoryPrduct'
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  models.Productlist:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  models.UpdateProductPut:
    properties:
      category_id:
        type: string
      description:
        type: string
      name:
        type: string
      photo:
        type: string
      price:
        type: number
    type: object
  models.UpdateUserSwag:
    properties:
      branch_id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /product:
    get:
      consumes:
      - application/json
      description: Get List product
      operationId: ProductPrimeryKey
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetProdctListBody
          schema:
            $ref: '#/definitions/models.GetListProductResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Product
      tags:
      - Product_5
    post:
      consumes:
      - application/json
      description: CreateProduct
      operationId: CreateProduct
      parameters:
      - description: CreateProductRequestBody
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.CreateProduct'
      produces:
      - application/json
      responses:
        "201":
          description: GetProductBody
          schema:
            $ref: '#/definitions/models.CreateProduct'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: CreateProduct
      tags:
      - Product_5
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Product
      operationId: DeleteProduct
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteProductBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Product
      tags:
      - Product_5
    get:
      consumes:
      - application/json
      description: GetByID Product
      operationId: Get_By_IDProduct
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: GetByIDProductBody
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: GetByID Product
      tags:
      - Product_5
    put:
      consumes:
      - application/json
      description: Update Product
      operationId: UpdateProduct
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateProductRequestBody
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProductPut'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateProductBody
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Product
      tags:
      - Product_5
  /user:
    get:
      consumes:
      - application/json
      description: Get List user
      operationId: UserPrimeryKey
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetUserListBody
          schema:
            $ref: '#/definitions/models.GetListUserResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List User
      tags:
      - User_2
    post:
      consumes:
      - application/json
      description: CreateUser
      operationId: CreateUser
      parameters:
      - description: CreateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserSwag'
      produces:
      - application/json
      responses:
        "201":
          description: GetUsertBody
          schema:
            $ref: '#/definitions/models.CreateUser'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: CreateUser
      tags:
      - User_2
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User
      operationId: DeleteUser
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteUserBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete User
      tags:
      - User_2
    get:
      consumes:
      - application/json
      description: GetByID User
      operationId: Get_By_IDUser
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: GetByIDUserBody
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: GetByID User
      tags:
      - User_2
    put:
      consumes:
      - application/json
      description: Update User
      operationId: UpdateUser
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateUserBody
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update User
      tags:
      - User_2
swagger: "2.0"